actuator/requirements-dev.txt
# requirements-dev.txt

black
darglint
mypy
pytest
ruff

# For testing CAN
python-can
pyserial
